common: Move findEvent to TransactionManager,use in allocation resolvers
