+++ packages/indexer-common/src/indexer-management/allocations.ts
@@ -42,10 +42,6 @@ import {
 
 import { BytesLike } from '@ethersproject/bytes'
 import pMap from 'p-map'
-import intersection from 'lodash.intersection'
-
-import updatedStakingAbi from '../abi/stakingUpdatedABI.json'
-const updatedStakingIface = new utils.Interface(updatedStakingAbi)
 
 export interface AllocateTransactionParams {
   indexer: string
@@ -190,63 +186,6 @@ export class AllocationManager {
       },
       { stopOnError: false },
     )
-  }
-
-  findEvent(
-    eventType: string,
-    contractInterface: utils.Interface,
-    logKey: string,
-    logValue: string,
-    receipt: ContractReceipt,
-  ): utils.Result | undefined {
-    const events: Event[] | providers.Log[] = receipt.events || receipt.logs
-    const decodedEvents: utils.Result[] = []
-
-    // With exponential rebates, the AllocationClosed event changed signature
-    // so it has a different topic. Until these changes are deployed in both mainnet and
-    // testnet, we need to search for both.
-    // TODO: Update to a new common-ts and remove this hack once exponential rebates is on mainnet
-    const newAbiTopic = updatedStakingIface.getEventTopic('AllocationClosed')
-    const expectedTopics = [contractInterface.getEventTopic(eventType)]
-    if (eventType == 'AllocationClosed') {
-      expectedTopics.push(newAbiTopic)
-    }
-
-    const result = events
-      .filter((event) => intersection(event.topics, expectedTopics).length > 0)
-      .map((event) => {
-        let decoded: utils.Result
-        if (eventType == 'AllocationClosed' && event.topics.includes(newAbiTopic)) {
-          decoded = updatedStakingIface.decodeEventLog(
-            eventType,
-            event.data,
-            event.topics,
-          )
-        } else {
-          decoded = contractInterface.decodeEventLog(eventType, event.data, event.topics)
-        }
-        decodedEvents.push(decoded)
-        return decoded
-      })
-      .find(
-        (eventLogs: utils.Result) =>
-          eventLogs[logKey].toLocaleLowerCase() === logValue.toLocaleLowerCase(),
-      )
-
-    this.logger.trace('Searched for event logs', {
-      function: 'findEvent',
-      expectedTopics,
-      events,
-      decodedEvents,
-      eventType,
-      logKey,
-      logValue,
-      receipt,
-      found: !!result,
-      result,
-    })
-
-    return result
   }
 
   async confirmActionExecution(
@@ -469,12 +408,13 @@ export class AllocationManager {
       )
     }
 
-    const createAllocationEventLogs = this.findEvent(
+    const createAllocationEventLogs = this.network.transactionManager.findEvent(
       'AllocationCreated',
       this.network.contracts.staking.interface,
       'subgraphDeploymentID',
       subgraphDeployment.bytes32,
       receipt,
+      this.logger,
     )
 
     if (!createAllocationEventLogs) {
@@ -603,12 +543,13 @@ export class AllocationManager {
       )
     }
 
-    const closeAllocationEventLogs = this.findEvent(
+    const closeAllocationEventLogs = this.network.transactionManager.findEvent(
       'AllocationClosed',
       this.network.contracts.staking.interface,
       'allocationID',
       allocationID,
       receipt,
+      this.logger,
     )
 
     if (!closeAllocationEventLogs) {
@@ -618,12 +559,13 @@ export class AllocationManager {
       )
     }
 
-    const rewardsEventLogs = this.findEvent(
+    const rewardsEventLogs = this.network.transactionManager.findEvent(
       'RewardsAssigned',
       this.network.contracts.rewardsManager.interface,
       'allocationID',
       allocationID,
       receipt,
+      this.logger,
     )
 
     const rewardsAssigned = rewardsEventLogs ? rewardsEventLogs.amount : 0
@@ -868,12 +810,13 @@ export class AllocationManager {
       )
     }
 
-    const closeAllocationEventLogs = this.findEvent(
+    const closeAllocationEventLogs = this.network.transactionManager.findEvent(
       'AllocationClosed',
       this.network.contracts.staking.interface,
       'allocationID',
       allocationID,
       receipt,
+      this.logger,
     )
 
     if (!closeAllocationEventLogs) {
@@ -883,12 +826,13 @@ export class AllocationManager {
       )
     }
 
-    const createAllocationEventLogs = this.findEvent(
+    const createAllocationEventLogs = this.network.transactionManager.findEvent(
       'AllocationCreated',
       this.network.contracts.staking.interface,
       'subgraphDeploymentID',
       closeAllocationEventLogs.subgraphDeploymentID,
       receipt,
+      this.logger,
     )
 
     if (!createAllocationEventLogs) {
@@ -898,12 +842,13 @@ export class AllocationManager {
       )
     }
 
-    const rewardsEventLogs = this.findEvent(
+    const rewardsEventLogs = this.network.transactionManager.findEvent(
       'RewardsAssigned',
       this.network.contracts.rewardsManager.interface,
       'allocationID',
       allocationID,
       receipt,
+      this.logger,
     )
 
     const rewardsAssigned = rewardsEventLogs ? rewardsEventLogs.amount : 0
